Partially implement execute/halt
git-svn-id: https://subversion.ews.illinois.edu/svn/fa16-ece391/_projects/loliOS@10727 0df279af-beec-47f8-8c4d-85a890efaaa8

+++ student-distrib/process.h
@@ -61,7 +61,7 @@ typedef struct {
 pcb_t *get_executing_pcb(void);
 
 /* Process syscall handlers */
-int32_t process_halt(uint8_t status);
+int32_t process_halt(uint32_t status);
 int32_t process_execute(const uint8_t *command);
 int32_t process_getargs(uint8_t *buf, int32_t nbytes);
